+++ src/main/java/me/drex/vanish/mixin/predicate/EntityFlagsFlagsPredicateMixin.java
@@ -10,9 +10,11 @@ import me.drex.vanish.api.VanishAPI;
 import me.drex.vanish.util.predicate.IEntityFlagsPredicate;
 //? if >= 26.2 {
 import net.minecraft.advancements.predicates.entity.EntityFlagsPredicate;
-//? } else {
+//? } else if >= 1.21.11 {
 /*import net.minecraft.advancements.criterion.EntityFlagsPredicate;
-*///? }
+*///? } else {
+//import net.minecraft.advancements.critereon.EntityFlagsPredicate;
+//? }
 import net.minecraft.world.entity.Entity;
 import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
@@ -50,9 +52,11 @@ public abstract class EntityFlagsFlagsPredicateMixin implements IEntityFlagsPred
             value = "FIELD",
             //? if >= 26.2 {
             target = "Lnet/minecraft/advancements/predicates/entity/EntityFlagsPredicate;CODEC:Lcom/mojang/serialization/Codec;",
-            //? } else {
-            /*target = "Lnet/minecraft/advancements/critereon/EntityFlagsPredicate;CODEC:Lcom/mojang/serialization/Codec;",
-            *///? }
+            //? } else if >= 1.21.11 {
+            /*target = "Lnet/minecraft/advancements/criterion/EntityFlagsPredicate;CODEC:Lcom/mojang/serialization/Codec;",
+            *///? } else {
+            //target = "Lnet/minecraft/advancements/critereon/EntityFlagsPredicate;CODEC:Lcom/mojang/serialization/Codec;",
+            //? }
             opcode = Opcodes.PUTSTATIC
         )
     )